+++ src/Plugins/PublishPressPro.php
@@ -9,7 +9,6 @@ namespace Junaidbhura\Composer\WPProPlugins\Plugins;
 
 use Junaidbhura\Composer\WPProPlugins\Http;
 use InvalidArgumentException;
-use UnexpectedValueException;
 
 /**
  * PublishPressPro class.
@@ -27,13 +26,12 @@ class PublishPressPro extends AbstractEddPlugin {
 	}
 
 	/**
-	 * Get the download URL for this plugin.
+	 * Get the download URL for this plugin from its API.
 	 *
 	 * @throws InvalidArgumentException If the package is unsupported.
-	 * @throws UnexpectedValueException If the response is invalid.
 	 * @return string
 	 */
-	public function getDownloadUrl() {
+	protected function getDownloadUrlFromApi() {
 		$id  = 0;
 		$env = null;
 		/**
@@ -106,23 +104,15 @@ class PublishPressPro extends AbstractEddPlugin {
 			$url     = ( getenv( "PUBLISHPRESS_{$env}_PRO_URL" ) ?: $url );
 		}
 
-		$http     = new Http();
-		$response = json_decode( $http->get( 'https://publishpress.com', array(
+		$http = new Http();
+
+		return $http->get( 'https://publishpress.com', array(
 			'edd_action' => 'get_version',
 			'license'    => $license,
 			'item_id'    => $id,
 			'url'        => $url,
 			'version'    => $this->version,
-		) ), true );
-
-		if ( ! is_array( $response ) ) {
-			throw new UnexpectedValueException( sprintf(
-				'Expected a JSON object from API for package %s',
-				'junaidbhura/' . $this->slug
-			) );
-		}
-
-		return $this->extractDownloadUrl( $response );
+		) );
 	}
 
 }
